Manage scopes of teamamte
I had initially designed the specification to default to user.profile.read when scopes for a teammate were not specified, but It enables users to configure it themselves.

However, as the current specification for setting values in scopes is unclear and could lead to errors, we have included ignore_changes for scopes in the samples, making them exempt from modifications.