fix: unexpected new value: .scopes: was cty.SetValEmpty(cty.String), but now null in teammate import
In teammate Import method,

After specifying .scopes = [] and applying, the following error occurs.

````
unexpected new value: .scopes: was cty.SetEmpty(cty.String), but now null.
````

This is because scopes is defined as `types.Set` type, so if it is not specified, it is specified as the default value.

To avoid the above, define .scopes with `[]types.String` type,
Specify the default with `[]types.String{}`.