Support SSO teammate
Provide a resource to manage SSO teammate.

Unlike the password teammate resource, this resource is specifically designed for SSO teammates.

I considered allowing the transition from a password teammate to an SSO teammate using a parameter such as is_sso. However, this approach would complicate the code and make it harder to maintain. Additionally, the handling of certain parameters differs between the two types: for instance, the first_name and last_name parameters are mandatory for SSO teammates but cannot be specified for password teammates.

For these reasons, I decided to implement SSO teammates as a separate resource.